backend/requirements.txt
# ========================================
# Talky.ai Backend - Python Dependencies
# Generated for Ubuntu (migrated from Windows)
# ========================================

# --- Core Framework ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
starlette==0.35.1
python-multipart==0.0.6
slowapi>=0.1.9


# --- Configuration & Settings ---
pydantic==2.12.5
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyYAML==6.0.1

# --- Database ---
sqlalchemy==2.0.25
psycopg2-binary==2.9.11
asyncpg==0.29.0
# supabase removed — using local PostgreSQL via asyncpg

# --- Cache & Queue ---
redis==5.0.1
hiredis==2.3.2

# --- HTTP Clients ---
httpx==0.27.2
aiohttp==3.13.2
aiofiles==23.2.1

# --- WebSocket ---
websockets==13.1

# --- AI Providers: STT ---
deepgram-sdk==5.3.0

# --- AI Providers: LLM ---
groq==0.37.0

# --- AI Providers: TTS ---
cartesia==2.0.17
google-cloud-texttospeech>=2.16.0

# --- Telephony ---
vonage>=3.0.0

# --- Authentication & Security ---
PyJWT>=2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
cryptography==46.0.3

# --- AI Assistant Agent ---
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.2.0

# --- Template Engine ---
Jinja2>=3.1.0

# --- Audio Processing ---
numpy>=1.26.0
pydub==0.25.1

# --- Monitoring & Utilities ---
prometheus-client==0.19.0
psutil>=5.9.0
pytz==2025.2


